fix: 161120801 Sentry Error: too many indices
In the case of a single-category categorical array (and perhaps other
single-element variables), the ndarray returned by `slice.as_array()` is
*reshaped* to remove a dimension. This unfortunately does not change the
return value of `CubeSlice.dimensions`, which will still report the
"squashed" dimension.

Calculate "reshaped" dimensions and use those for calculating
`.valid_indices()` so the shape of the indices matches the value array.