rfctr: extract CrunchCube._raw_cube_array()
This functionality of mapping the flat values from the cube response to
an ndarray of the right shape is very coherent and a good candidate for
extracting to a collaborator object. Extracting it (for now) to
a separate method clarifies the intent of ._as_array() and paves the way
for further extraction to a _Measure object.