Enable correct pruning for insertions
* Insertions are _almost never_ pruned. The only time they're pruned is
when their base counts would all be zero. That can consequently be
regarded as - all base counts of the opposing dimension vectors are all
zeros. We need to prune only in this case (because the opposing vectors
will _all_ be pruned, and there will be nothing left in the insertions
vectors to display.
* Expand the API of the Means slices, to support row and col iteration
(mainly to support getting the labels from cr.exporter.