fix: IndexError arrays used as indices must be int
Exception (from numpy) was:
IndexError: arrays used as indices must be of integer (or boolean) type.

The root cause was the indexing (slicing) array for row ordering was
defaulting to dtype=float, but only when empty, such that the data type
could not be inferred from array values. The problematic analysis had
all row elements marked as missing, so the valid-element-idxs collection
was empty.

Add integration and unit tests, and fix for columns as well as rows.