rfctr: extract _StrandInsertions
Do a general overhaul on insertions in the process. Next step is to
remove separate _MatrixInsertions and _StrandInsertions objects and let
MatrixWithInsertions and StrandWithInsertions take care of that
themselves. They already have access to the dimensions, so there's no
compelling reason to have a separate top-level object do that work.