rfctr: extract StripeColumnVector
Remove use of matrix vectors for stripe columns. It turns out only
a small fraction of what they do is required in that role and what is
there can probably be removed here shortly. A stripe doesn't actually
have any columns, so the current thrust is to actually get rid of those
and only leave row vectors.