rfctr: mv row/column_margin to rows_/columns_margin
Rename interface properties `row_margin` and `column_margin` to
`rows_margin` and `columns_margin`.

Rationale: row-margin is a thing, it applies to a single row and is
a scalar value. rows_margin (the thing in question here) applies to *all
the rows in a rows_dimension* and is an array value of length row_count.