rfctr: Make CubePartition the partition base-class
There are very few things that are common to all cube-partition types,
but dimensions and those properties derived from that are common.

Expand the responsibility of the existing CubePartition beyond the
partition factory role to also provide the dimension aspects of the
interface. I expect the common set will grow to include things like
`.has_scale_means` and probably `.has_means`.