rfctr: improve .scale_mean_pairwise_indices return type
Avoid using numpy types as interface (return) types. Use a tuple for the
container of the column-idx tuples instead of an np.ndarray.

Using a Python-native type makes using the method easier, requiring no
special methods the user may not be as familiar with. Using an immutable
type in a lazyproperty is wise because it prevents a caller from
mutating a value that may be returned to a different caller in another
call.