fix: pairwise_indices array collapsing when empty
The array returned by
`_ColumnPairwiseSignificance.pairwise_indices()` loses its shape when it
contains only empty tuples, becoming [] instead of for example
`[[(), ()], [(), ()]]`. This causes problems for exporter which relies
on the shape being the same as `.counts` (or other measures).

Explicitly define the shape and type of the array to avoid this
automatic numpy behavior.