rfctr: rename _BaseInsertedVector
The "Matrix" specifier is redundant, this class is local to the `matrix`
module and so can only apply here. I think this naming came from early
on when we though that a stripe would just be an unopposed slice-vector,
but that of course turned out not to be the case.