mtx: add Assembler.rows_dimension_numeric_values
Also, remove `.rows_dimension_numeric_values` from `_Slice` public
interface as it was being used for internal purposes (by
`cr.cube.measures.pairwise_significance`). It's poor form to extend a
public interface for private (internals) purposes. Pairwise significance
should get this value from assembler or wherever is most convenient, not
from _Slice (the interface object). This is likely to be remedied when
measures are extracted to a distinct layer/object-hierarchy and have
access to the cube-result matrix.