xfail: add XFAILs for _UnweightedCubeCounts class
The next step is to extract the unweighted cube-counts as a separate
cube-measure. This places all distinct cube-measures on an equal
footing. Right now that will be UnweightedCubeCounts,
WeightedCubeCounts, and Means, but may extend to StdDevs and
Covariances, etc.

It's important for them to be distinct, because the presence of one does
not imply the presence of another, like just because unweighted
cube-counts are present doesn't mean that weighted-cube-counts are.