fix: Python 2.7 mock incompatibility with classmethod
Early versions of `mock` used in Python 2.7 did not behave well with
autospec'ing of classmethods. Disable autospec on a couple tests that
pass fine in Python 3. The cost is that the unit test will no longer
recognize parameter count mismatches, so not a terribly big deal. That
sort of change will produce other test failures, it just won't pinpoint
the problem quite as nicely.