fix: sort-by-value keyword to "percent"
Measure keyword is distinct for univariate measure, "percent" instead of
"col_percent" as we use for the measure itself in exporter.

This makes sense because there is only one kind of percent measure for a
frequency (always univariate) analysis and that is not column-percent.
Rather it is what we might call "total_percent" but there's no need for
a modifier word in this case since there's just the one.