sort: wire-up sort-by-value for available measures
All sort-by-value types rely on a measure and can only be implemented
when that measure has consolidated (available in
`cr.cube.stripe.measures` or `cr.cube.matrix.measures`).

Implement sort-by-value for all measures that have been consolidated so
far. Note that no margins have been consolidated yet and so
sort-rows-by-marginal-column is not yet implemented (although it is
straightforward to do so in a way very similar to the ones so far).